+++ src/main/java/com/example/forummanagementsystem/exceptions/EntityNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.forummanagementsystem.exceptions;
 
 public class EntityNotFoundException extends RuntimeException {
 
-    public EntityNotFoundException(String type, int id) {
+    public EntityNotFoundException(String type, Long id) {
         this(type, "id", String.valueOf(id));
     }
