+++ src/main/java/com/example/forummanagementsystem/services/PostService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface PostService {
 
     List<Post> getAll();
-    Post getById(int id);
+    Post getById(Long id);
 
     Post getByTitle(String name);
 
     void create(Post post);
     void update(Post post);
-    void delete(int id);
+    void delete(Long id);
 
 }
